refactor: 폴더명 변경 (module -> webviewBridge)
